feat: users are linked with roles in the index, show and edit methods

+++ app/Http/Controllers/UserController.php
@@ -45,12 +45,15 @@ class UserController extends Controller
     {
         // $user=User::findOrFail($id);
         // dd($user);
+        $user->load('roles');
         return view('users.show', compact('user'));
     }
 
     public function edit(User $user)
     {
-        return view('users.edit', compact('user'));
+        $roles = Role::all()->pluck('name', 'id');
+        $user->load('roles');
+        return view('users.edit', compact('user', 'roles'));
     }
 
     public function update(UserEditRequest $request, User $user)
@@ -62,12 +65,21 @@ class UserController extends Controller
             $data['password'] = bcrypt($password);
         
         $user->update($data);
+
+        $roles = $request->input('roles', []);
+        $user->syncRoles($roles);
+
         return redirect()->route('users.show', $user->id)->with('success', 'Successful data update');
     }
 
     public function destroy(User $user)
     {
+        if (auth()->user()->id == $user->id) {
+            return redirect()->route('users.index');
+        }
+        
         $user->delete();
+
         return back()->with('success', 'Successfully deleted user');
     }
 }
